back down to 10.5 for OSX (10.4 still broken), fixed 64bit warning in tracer

+++ EngineRoom/tracer.c
@@ -477,10 +477,11 @@ tracerSymbolStoreShow( TRACER_SYMBOL_STORE *store )
 				  (void*)store,
 				  (unsigned long int)store->symbols_size,
 				  (unsigned long int)store->symbols_capacity,
-				  (unsigned long int)store->symbols_capacity * sizeof(*store->symbols) / 1024UL,
-				  (unsigned long int)store->strtab_capacity / 1024UL,
-				  ((unsigned long int)store->strtab_size + (unsigned long int)store->symbols_capacity * sizeof(*store->symbols)) / 1024UL);
-	
+				  (unsigned long int)store->symbols_capacity * (unsigned long int)sizeof(*store->symbols) / 1024UL,
+				  (unsigned long int) (store->strtab_capacity / 1024),
+				  ((unsigned long int)store->strtab_size + (unsigned long int)store->symbols_capacity * (unsigned long int)sizeof(*store->symbols)) / 1024UL
+				  );
+
 	return TRACER_SUCCESS;
 }
